Common way to access the data ptr
